codex: carve context_trace.rs (pure move)
Managed-context support machinery: request/context payload snapshots
from the trace archive, trace-bundle discovery + fingerprints, context
summaries, and usage extraction from traced inference payloads; 13
tests ride along. Snapshot/index struct fields widened pub(crate) for
the mod.rs consumers.
